+++ lib/segment/src/index/field_index/field_index_base/field_index.rs
@@ -3,30 +3,21 @@ use std::path::PathBuf;
 
 use common::counter::hardware_counter::HardwareCounterCell;
 use common::types::PointOffsetType;
-use itertools::Either;
 use serde_json::Value;
 
-use super::bool_index::BoolIndex;
-use super::bool_index::immutable_bool_index::ImmutableBoolIndexBuilder;
-use super::bool_index::mutable_bool_index::MutableBoolIndexBuilder;
-use super::facet_index::FacetIndexEnum;
-use super::full_text_index::mmap_text_index::FullTextMmapIndexBuilder;
-use super::full_text_index::text_index::{
-    FullTextGridstoreIndexBuilder, FullTextIndex, PayloadMatchQueryType,
-};
-use super::geo_index::{GeoMapIndexGridstoreBuilder, GeoMapIndexMmapBuilder};
-use super::map_index::{MapIndex, MapIndexGridstoreBuilder, MapIndexMmapBuilder};
-use super::null_index::immutable_null_index::ImmutableNullIndexBuilder;
-use super::numeric_index::{
-    NumericIndex, NumericIndexGridstoreBuilder, NumericIndexMmapBuilder, StreamRange,
-};
+use super::payload_field_index::PayloadFieldIndex;
+use super::value_indexer::ValueIndexer;
 use crate::common::Flusher;
 use crate::common::operation_error::OperationResult;
-use crate::data_types::order_by::OrderValue;
+use crate::index::field_index::bool_index::BoolIndex;
+use crate::index::field_index::facet_index::FacetIndexEnum;
+use crate::index::field_index::full_text_index::text_index::{
+    FullTextIndex, PayloadMatchQueryType,
+};
 use crate::index::field_index::geo_index::GeoMapIndex;
+use crate::index::field_index::map_index::MapIndex;
 use crate::index::field_index::null_index::NullIndex;
-use crate::index::field_index::null_index::mutable_null_index::MutableNullIndexBuilder;
-use crate::index::field_index::numeric_index::NumericIndexInner;
+use crate::index::field_index::numeric_index::{NumericFieldIndex, NumericIndex};
 use crate::index::field_index::{CardinalityEstimation, PayloadBlockCondition};
 use crate::index::payload_config::{
     FullPayloadIndexType, IndexMutability, PayloadIndexType, StorageType,
@@ -34,99 +25,8 @@ use crate::index::payload_config::{
 use crate::telemetry::PayloadIndexTelemetry;
 use crate::types::{
     DateTimePayloadType, FieldCondition, FloatPayloadType, IntPayloadType, Match, MatchPhrase,
-    MatchText, MatchTextAny, PayloadKeyType, RangeInterface, UuidIntType, UuidPayloadType,
+    MatchText, MatchTextAny, PayloadKeyType, UuidIntType, UuidPayloadType,
 };
-
-pub trait PayloadFieldIndex {
-    /// Return number of points with at least one value indexed in here
-    fn count_indexed_points(&self) -> usize;
-
-    /// Remove db content or files of the current payload index
-    fn wipe(self) -> OperationResult<()>;
-
-    /// Return function that flushes all pending updates to disk.
-    fn flusher(&self) -> Flusher;
-
-    fn files(&self) -> Vec<PathBuf>;
-
-    fn immutable_files(&self) -> Vec<PathBuf>;
-
-    /// Get iterator over points fitting given `condition`
-    /// Return `None` if condition does not match the index type
-    fn filter<'a>(
-        &'a self,
-        condition: &'a FieldCondition,
-        hw_counter: &'a HardwareCounterCell,
-    ) -> OperationResult<Option<Box<dyn Iterator<Item = PointOffsetType> + 'a>>>;
-
-    /// Return estimation of amount of points which satisfy given condition.
-    /// Returns `Ok(None)` if the condition does not match the index type
-    fn estimate_cardinality(
-        &self,
-        condition: &FieldCondition,
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<Option<CardinalityEstimation>>;
-
-    /// Iterate conditions for payload blocks with minimum size of `threshold`
-    /// Required for building HNSW index
-    fn for_each_payload_block(
-        &self,
-        threshold: usize,
-        key: PayloadKeyType,
-        f: &mut dyn FnMut(PayloadBlockCondition) -> OperationResult<()>,
-    ) -> OperationResult<()>;
-}
-
-pub trait ValueIndexer {
-    type ValueType;
-
-    /// Add multiple values associated with a single point
-    /// This function should be called only once for each point
-    fn add_many(
-        &mut self,
-        id: PointOffsetType,
-        values: Vec<Self::ValueType>,
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<()>;
-
-    /// Extract index-able value from payload `Value`
-    fn get_value(value: &Value) -> Option<Self::ValueType>;
-
-    /// Try to extract index-able values from payload `Value`, even if it is an array
-    fn get_values(value: &Value) -> Vec<Self::ValueType> {
-        match value {
-            Value::Array(values) => values.iter().filter_map(|x| Self::get_value(x)).collect(),
-            _ => Self::get_value(value).map(|x| vec![x]).unwrap_or_default(),
-        }
-    }
-
-    /// Add point with payload to index
-    fn add_point(
-        &mut self,
-        id: PointOffsetType,
-        payload: &[&Value],
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<()> {
-        self.remove_point(id)?;
-        let mut flatten_values: Vec<_> = vec![];
-        for value in payload {
-            match value {
-                Value::Array(values) => {
-                    flatten_values.extend(values.iter().filter_map(|x| Self::get_value(x)));
-                }
-                _ => {
-                    if let Some(x) = Self::get_value(value) {
-                        flatten_values.push(x);
-                    }
-                }
-            }
-        }
-        self.add_many(id, flatten_values, hw_counter)
-    }
-
-    /// remove a point from the index
-    fn remove_point(&mut self, id: PointOffsetType) -> OperationResult<()>;
-}
 
 /// Common interface for all possible types of field indexes
 /// Enables polymorphism on field indexes
@@ -547,231 +447,5 @@ impl FieldIndex {
             FieldIndex::UuidMapIndex(index) => index.get_storage_type(),
             FieldIndex::NullIndex(index) => index.get_storage_type(),
         }
-    }
-}
-
-/// Common interface for all index builders.
-pub trait FieldIndexBuilderTrait {
-    /// The resulting type of the index.
-    type FieldIndexType;
-
-    /// Start building the index, e.g. create a database column or a directory.
-    /// Expected to be called exactly once before any other method.
-    fn init(&mut self) -> OperationResult<()>;
-
-    fn add_point(
-        &mut self,
-        id: PointOffsetType,
-        payload: &[&Value],
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<()>;
-
-    fn finalize(self) -> OperationResult<Self::FieldIndexType>;
-
-    /// Create an empty index for testing purposes.
-    #[cfg(test)]
-    fn make_empty(mut self) -> OperationResult<Self::FieldIndexType>
-    where
-        Self: Sized,
-    {
-        self.init()?;
-        self.finalize()
-    }
-}
-
-/// Builders for all index types
-pub enum FieldIndexBuilder {
-    IntMmapIndex(NumericIndexMmapBuilder<IntPayloadType, IntPayloadType>),
-    IntGridstoreIndex(NumericIndexGridstoreBuilder<IntPayloadType, IntPayloadType>),
-    DatetimeMmapIndex(NumericIndexMmapBuilder<IntPayloadType, DateTimePayloadType>),
-    DatetimeGridstoreIndex(NumericIndexGridstoreBuilder<IntPayloadType, DateTimePayloadType>),
-    IntMapMmapIndex(MapIndexMmapBuilder<IntPayloadType>),
-    IntMapGridstoreIndex(MapIndexGridstoreBuilder<IntPayloadType>),
-    KeywordMmapIndex(MapIndexMmapBuilder<str>),
-    KeywordGridstoreIndex(MapIndexGridstoreBuilder<str>),
-    FloatMmapIndex(NumericIndexMmapBuilder<FloatPayloadType, FloatPayloadType>),
-    FloatGridstoreIndex(NumericIndexGridstoreBuilder<FloatPayloadType, FloatPayloadType>),
-    GeoMmapIndex(GeoMapIndexMmapBuilder),
-    GeoGridstoreIndex(GeoMapIndexGridstoreBuilder),
-    FullTextMmapIndex(FullTextMmapIndexBuilder),
-    FullTextGridstoreIndex(FullTextGridstoreIndexBuilder),
-    BoolMmapIndex(ImmutableBoolIndexBuilder),
-    BoolGridstoreIndex(MutableBoolIndexBuilder),
-    UuidMmapIndex(MapIndexMmapBuilder<UuidIntType>),
-    UuidGridstoreIndex(MapIndexGridstoreBuilder<UuidIntType>),
-    MutableNullIndex(MutableNullIndexBuilder),
-    ImmutableNullIndex(ImmutableNullIndexBuilder),
-}
-
-impl FieldIndexBuilderTrait for FieldIndexBuilder {
-    type FieldIndexType = FieldIndex;
-
-    fn init(&mut self) -> OperationResult<()> {
-        match self {
-            Self::IntMmapIndex(index) => index.init(),
-            Self::IntGridstoreIndex(index) => index.init(),
-            Self::DatetimeMmapIndex(index) => index.init(),
-            Self::DatetimeGridstoreIndex(index) => index.init(),
-            Self::IntMapMmapIndex(index) => index.init(),
-            Self::IntMapGridstoreIndex(index) => index.init(),
-            Self::KeywordMmapIndex(index) => index.init(),
-            Self::KeywordGridstoreIndex(index) => index.init(),
-            Self::FloatMmapIndex(index) => index.init(),
-            Self::FloatGridstoreIndex(index) => index.init(),
-            Self::GeoMmapIndex(index) => index.init(),
-            Self::GeoGridstoreIndex(index) => index.init(),
-            Self::BoolMmapIndex(index) => index.init(),
-            Self::BoolGridstoreIndex(index) => index.init(),
-            Self::FullTextMmapIndex(builder) => builder.init(),
-            Self::FullTextGridstoreIndex(builder) => builder.init(),
-            Self::UuidMmapIndex(index) => index.init(),
-            Self::UuidGridstoreIndex(index) => index.init(),
-            Self::MutableNullIndex(index) => index.init(),
-            Self::ImmutableNullIndex(index) => index.init(),
-        }
-    }
-
-    fn add_point(
-        &mut self,
-        id: PointOffsetType,
-        payload: &[&Value],
-        hw_counter: &HardwareCounterCell,
-    ) -> OperationResult<()> {
-        match self {
-            Self::IntMmapIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::IntGridstoreIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::DatetimeMmapIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::DatetimeGridstoreIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::IntMapMmapIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::IntMapGridstoreIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::KeywordMmapIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::KeywordGridstoreIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::FloatMmapIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::FloatGridstoreIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::GeoMmapIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::GeoGridstoreIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::BoolGridstoreIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::BoolMmapIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::FullTextMmapIndex(builder) => {
-                FieldIndexBuilderTrait::add_point(builder, id, payload, hw_counter)
-            }
-            Self::FullTextGridstoreIndex(builder) => {
-                FieldIndexBuilderTrait::add_point(builder, id, payload, hw_counter)
-            }
-            Self::UuidMmapIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::UuidGridstoreIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::MutableNullIndex(index) => index.add_point(id, payload, hw_counter),
-            Self::ImmutableNullIndex(index) => index.add_point(id, payload, hw_counter),
-        }
-    }
-
-    fn finalize(self) -> OperationResult<FieldIndex> {
-        Ok(match self {
-            Self::IntMmapIndex(index) => FieldIndex::IntIndex(index.finalize()?),
-            Self::IntGridstoreIndex(index) => FieldIndex::IntIndex(index.finalize()?),
-            Self::DatetimeMmapIndex(index) => FieldIndex::DatetimeIndex(index.finalize()?),
-            Self::DatetimeGridstoreIndex(index) => FieldIndex::DatetimeIndex(index.finalize()?),
-            Self::IntMapMmapIndex(index) => FieldIndex::IntMapIndex(index.finalize()?),
-            Self::IntMapGridstoreIndex(index) => FieldIndex::IntMapIndex(index.finalize()?),
-            Self::KeywordMmapIndex(index) => FieldIndex::KeywordIndex(index.finalize()?),
-            Self::KeywordGridstoreIndex(index) => FieldIndex::KeywordIndex(index.finalize()?),
-            Self::FloatMmapIndex(index) => FieldIndex::FloatIndex(index.finalize()?),
-            Self::FloatGridstoreIndex(index) => FieldIndex::FloatIndex(index.finalize()?),
-            Self::GeoMmapIndex(index) => FieldIndex::GeoIndex(index.finalize()?),
-            Self::GeoGridstoreIndex(index) => FieldIndex::GeoIndex(index.finalize()?),
-            Self::BoolGridstoreIndex(index) => {
-                FieldIndex::BoolIndex(BoolIndex::from(index.finalize()?))
-            }
-            Self::BoolMmapIndex(index) => FieldIndex::BoolIndex(BoolIndex::from(index.finalize()?)),
-            Self::FullTextMmapIndex(builder) => FieldIndex::FullTextIndex(builder.finalize()?),
-            Self::FullTextGridstoreIndex(builder) => FieldIndex::FullTextIndex(builder.finalize()?),
-            Self::UuidMmapIndex(index) => FieldIndex::UuidMapIndex(index.finalize()?),
-            Self::UuidGridstoreIndex(index) => FieldIndex::UuidMapIndex(index.finalize()?),
-            Self::MutableNullIndex(index) => {
-                FieldIndex::NullIndex(NullIndex::from(index.finalize()?))
-            }
-            Self::ImmutableNullIndex(index) => {
-                FieldIndex::NullIndex(NullIndex::from(index.finalize()?))
-            }
-        })
-    }
-}
-
-pub enum NumericFieldIndex<'a> {
-    IntIndex(&'a NumericIndexInner<IntPayloadType>),
-    FloatIndex(&'a NumericIndexInner<FloatPayloadType>),
-}
-
-impl<'a> StreamRange<OrderValue> for NumericFieldIndex<'a> {
-    fn stream_range(
-        &self,
-        range: &RangeInterface,
-    ) -> OperationResult<impl DoubleEndedIterator<Item = (OrderValue, PointOffsetType)> + '_> {
-        Ok(match self {
-            NumericFieldIndex::IntIndex(index) => Either::Left(
-                index
-                    .stream_range(range)?
-                    .map(|(v, p)| (OrderValue::from(v), p)),
-            ),
-            NumericFieldIndex::FloatIndex(index) => Either::Right(
-                index
-                    .stream_range(range)?
-                    .map(|(v, p)| (OrderValue::from(v), p)),
-            ),
-        })
-    }
-}
-
-impl<'a> NumericFieldIndex<'a> {
-    pub fn get_ordering_values(
-        &self,
-        idx: PointOffsetType,
-    ) -> impl Iterator<Item = OrderValue> + 'a {
-        match self {
-            NumericFieldIndex::IntIndex(index) => Either::Left(
-                index
-                    .get_values(idx)
-                    .into_iter()
-                    .flatten()
-                    .map(OrderValue::Int),
-            ),
-            NumericFieldIndex::FloatIndex(index) => Either::Right(
-                index
-                    .get_values(idx)
-                    .into_iter()
-                    .flatten()
-                    .map(OrderValue::Float),
-            ),
-        }
-    }
-}
-
-/// Read-only abstraction over a per-key numeric index.
-///
-/// Implemented by the appendable [`NumericFieldIndex`] today; a future
-/// `ReadOnlySegment` will provide its own concrete numeric-index type with
-/// the same shape, so order-by reads can share one implementation.
-///
-/// Returned iterators borrow from `&self` — callers hold them within the
-/// scope of the borrow.
-pub trait NumericFieldIndexRead {
-    fn get_ordering_values(&self, idx: PointOffsetType) -> impl Iterator<Item = OrderValue> + '_;
-
-    fn stream_range(
-        &self,
-        range: &RangeInterface,
-    ) -> OperationResult<impl DoubleEndedIterator<Item = (OrderValue, PointOffsetType)> + '_>;
-}
-
-impl<'a> NumericFieldIndexRead for NumericFieldIndex<'a> {
-    fn get_ordering_values(&self, idx: PointOffsetType) -> impl Iterator<Item = OrderValue> + '_ {
-        NumericFieldIndex::get_ordering_values(self, idx)
-    }
-
-    fn stream_range(
-        &self,
-        range: &RangeInterface,
-    ) -> OperationResult<impl DoubleEndedIterator<Item = (OrderValue, PointOffsetType)> + '_> {
-        StreamRange::stream_range(self, range)
     }
 }
