+++ lib/segment/src/index/field_index/field_index_base/mod.rs
@@ -1,0 +1,9 @@
+mod builder;
+mod field_index;
+mod payload_field_index;
+mod value_indexer;
+
+pub use builder::{FieldIndexBuilder, FieldIndexBuilderTrait};
+pub use field_index::FieldIndex;
+pub use payload_field_index::PayloadFieldIndex;
+pub use value_indexer::ValueIndexer;
